chore: reduces circular dependencies with some cleanup
